vectorcam-refsr/container/code/requirements.txt
torch
torchvision
numpy
pillow
opencv-python
imageio
boto3